requirements.txt
# Core dependencies
rich>=13.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai-agents[littlellm]>=0.0.16
typer>=0.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional dependencies for different AI providers
anthropic>=0.8.0
openai>=1.0.0
